EY-Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
httpx
websockets
asyncpg
google-generativeai
faker
dateparser

